# [WIP] Refactor constraint transformations + ELBO maximization