--- a/src/Celeste.jl
+++ b/src/Celeste.jl
@@ -9,6 +9,7 @@ include("SensitiveFloats.jl")
 
 include("Model.jl")
 include("Infer.jl")
+include("ConstraintTransforms.jl")
 include("Transform.jl")
 include("PSF.jl")
 include("SDSSIO.jl")