--- a/src/Model.jl
+++ b/src/Model.jl
@@ -5,8 +5,7 @@ using StaticArrays
 # parameter types
 export Image, SkyPatch, PsfComponent,
        GalaxyComponent, GalaxyPrototype,
-       PriorParams, UnconstrainedParams,
-       CanonicalParams, BrightnessParams, StarPosParams,
+       PriorParams, CanonicalParams, BrightnessParams, StarPosParams,
        GalaxyPosParams, GalaxyShapeParams,
        PsfParams, RawPSF, CatalogEntry,
        ParamSet
@@ -20,8 +19,8 @@ export band_letters, D, Ia, B,
        shape_standard_alignment,
        brightness_standard_alignment,
        gal_shape_alignment,
-       ids_names, ids_free_names,
-       ids, ids_free, star_ids, gal_ids, gal_shape_ids, psf_ids, bids
+       ids_names,
+       ids, star_ids, gal_ids, gal_shape_ids, psf_ids, bids
 
 
 import Base.convert