--- a/src/deterministic_vi/elbo_args.jl
+++ b/src/deterministic_vi/elbo_args.jl
@@ -197,7 +197,7 @@ type ElboArgs{NumType <: Number}
     # If this is set to Inf, the bivariate normals will be evaluated at all
     # points irrespective of their distance from the mean.
     num_allowed_sd::Float64
-    
+
     # If true, only render star parameters for active sources.
     active_source_star_only::Bool