… src/Transform.jl
… src/deterministic_vi/NewtonMaximize.jl
… src/deterministic_vi/elbo_objective.jl
… src/deterministic_vi/maximize_elbo.jl
… src/joint_infer.jl
… src/model/param_set.jl
… test/SampleData.jl
… test/runtests.jl
… test/test_constraints.jl
… test/test_derivatives.jl
… test/test_elbo.jl
… test/test_joint_infer.jl
… test/test_optimization.jl
… test/test_transforms.jl