There's still quite a bit of work to be done here, but I wanted to go ahead and open this PR so that it gets some exposure to the other maintainers. It might be quite a slog towards the finish line. I've added the new code to the source tree so that it can be reviewed, but haven't "hooked it up" to the other code yet (e.g. the old code is still being used on this branch). 

Following @jeff-regier's advice, this PR will ignore `PSF.jl` (and thus will keep the old `Transform.jl` around so `PSF.jl` can use it).

This refactors the constraint transformation code (previously in `Transform.jl`, now in `ConstraintTransforms.jl`) to reduce allocations, make things type-stable, and use AD for the derivatives. IMO, this new design should also be much easier to extend/modify in the future than the old design. This additionally refactors `maximize_elbo.jl` to use the new `ConstraintTransforms.jl` code.

I should write up a more detailed description of the changes at some point, but feel free to ask questions in the meantime.

There are a number of things left to do:

- [x] explore/replace `enforce_bounds`
- [x] allow callers to pass in their own constrain/unconstrain functions (and provide an equivalent of the default transforms given by `get_mp_transform` in the old code)
- [x] default constrain/unconstrain functions can optionally accept pre-allocated location constraints
- [x] add replacement for `omitted_ids`
- [x] write tests for new code and fix the inevitable bugs that will crop up
    - [x] test ConstraintTransforms.jl
    - [x] rewrite existing optimization tests to use NewtonMaximize
    - [x] remove/fix any other broken tests 
- [x] refactor callers of `maximize_f`:
    - [x] `infer_source` in `DeterministicVI.jl`
    - [x] `one_node_joint_infer` in `joint_infer.jl`
    - [x] `infer_source_fft` in `DeterministicVIImagePSF.jl`
- [x] run benchmarks in `benchmarks` folder once the new code is "hooked up" (I've done piecemeal benchmarks and the performance looks good so far)

Closes #492.