@jeff-regier 2017-01-11
It looks to me like `enforce_bounds!` is just used to initialize the variational parameters. But shouldn't they already be in bounds, if they're initialized with [`generic_init_source `](https://github.com/jeff-regier/Celeste.jl/blob/master/src/DeterministicVI.jl#L27) or [`catalog_init_source`](https://github.com/jeff-regier/Celeste.jl/blob/master/src/DeterministicVI.jl#L49)?

@rgiordan -- Would you explain the idea behind `explore_bounds!`? If it's just for unit tests, I'd rather move it to the `tests/` directory, and explicitly call it in the tests before calling `maximize_f`.