@rgiordan 2017-01-16
```enforce_bounds!``` ensures that the program will not terminate if you pass parameters to the optimization that cannot be expressed within the transform bounds.

Here's an example of why that's useful outside of tests.  Through ```simplex_min``` in ```Transform.jl```, we don't allow E[a] to be less than ```0.005```.  Suppose I want to start optimization at essentially "all star".  If I actually initialize to ```a[1] = 1```, Celeste will crash when it is unable to express this in constrained space.  To avoid having to look up the legal minimum, you can just set the initial parameters to what you want and guarantee that Celeste will not crash.

As another use case, I don't think we want Celeste to crash if there is an initial catalog value that cannot be expressed in the constrained space.  Rather, we'd like it to be brought as close as is allowed and proceed with optimization.