@jeff-regier 2017-01-16
I think the easiest way to get this PR merged is to keep `enforce_bounds!` around, effectively re-implementing it to do what the old version does. I don't entirely agree with what Ryan says about its usefulness outside of tests:  `catalog_init_source` already does thresholding, so it should give reasonable, in-bounds values for initialization, regardless of the current catalog. (In the future, we may want `catalog_init_source` to also return the constraints for a particular light source.) Second, nowhere outside of the tests are parameters set to values that could be out of bounds, e.g. `a[1] = 1` doesn't happen outside of tests. That said, it is easiest not to change too many aspects of the code at once, and `enforce_bounds!` doesn't seem like a problem now that I understand it, just a particular design decision.