@jrevels 2017-01-16
I'm fine with reimplementing `enforce_bounds!`, I'll play around with that tomorrow.