@rgiordan 2017-01-16
On Mon, Jan 16, 2017 at 1:46 PM, Jeffrey Regier <notifications@github.com>
wrote:

> e.g. a[1] = 1 doesn't happen outside of tests.
>

I do this all the time when debugging the model or using it interactively.