@jeff-regier 2017-01-18
@jrevels Would add an equality constraint type too? Something like

```
immutable EqualityConstraint <: Constraint
    c::Vector{Float64}
end
```

For a `ParameterConstraint{EqualityConstraint}` instance, I think `free_indices` would just be `Int[]`---the bound parameters are omitted from unconstrained parameterization. When mapping from the unconstrained parameters to the constrained parameters, the entries of `c` are inserted. So `length(c) == length(bound_indices)`, if I understand your approach correctly.

@rgiordan -- Will this functionality suffice for issue #480?