@jrevels 2017-01-18
Adding that should be pretty easy. Is there any other code I'd need to update to use that, or will it mainly be useful in the future/for debugging?