@jeff-regier 2017-01-18
Great, thanks Jarrett. It's just for debugging, at least for now. Ryan is having trouble getting to the bottom of some problems with the FFT optimization without it.