@rgiordan 2017-01-18
@jeff-regier , I'm not sure what you're proposing re: https://github.com/jeff-regier/Celeste.jl/issues/480.  Are you suggesting adding the capability to specify a custom ```constrain_model``` function when you call ```newton_maximize!```?  Wouldn't it be simpler to just fix ```omitted_ids```?