@jeff-regier 2017-01-18
Yes, that's the idea. It should be easier to add and maintain an `EqualityConstraint` type than to add `omitted_ids` logic back.

As a stopgap, could you set the bounds for the parameters you don't want to change to a very narrow range? Would the let you continue to make progress while Jarrett works on Transform?