@rgiordan 2017-01-18
Tight bounds will result in large gradients and Hessians for those parameters.

I don't quite understand your reluctance to fix ```omitted_ids```.  It looks like someone even tried to implement it (e.g., there is an ```all_kept_ids``` defined), but they didn't test it and it has a (probably minor) bug.  Please fix ```omitted_ids```.