@jrevels 2017-01-18
@jeff-regier I've updated this PR's TODO list, can you review what I've put there and let me know if I missed anything/should change anything? 

I finished a couple of the points we talked about in today's meeting. If I can keep going at this pace (i.e. the testing/bug-fixing doesn't reveal any monstrous problems), then hopefully this PR can make it in before the January hackathon.