@jeff-regier 2017-01-18
@jrevels -- That TODO list looks good to me. Please also include a unit test that demonstrates how to use `EqualityConstraint` to optimize just the star parameters, while leaving the other parameters fixed at their initial value. Thanks for working on this issue.