@jeff-regier 2017-02-20
All tests pass for me after rebasing on master.

Should this whole PR get squashed into one commit? Or do you want to group it into multiple commits?