@jeff-regier 2017-02-20
Scores look good from `validate_on_stripe82.jl --joint`! I think we're in business.