@jeff-regier 2017-02-20
The performance regression looks real, unfortunately, at least with Julia 0.5. From `benchmark_sixteenth_degree.jl`, with 3 threads, on a macintosh:

On master:
```
386.197440 seconds (1.04 G allocations: 63.712 GB, 4.50% gc time)
```

On "jr/constraints" (following rebase)
```
477.714822 seconds (1.05 G allocations: 52.766 GB, 4.10% gc time)
```