@jrevels 2017-02-20
Hmm...benchmarking `NewtonMaximize.maximize!` vs. the old `maximize_f` with a single source and `DeterministicVI.elbo` as the objective indicates that NewtonMaximize is slightly faster with a fairly substantial (~40%) memory improvement (on v0.6 anyway). I would benchmark with multiple simultaneous sources, but the old Transform code errors out in that case.

This PR shouldn't cause drastic performance changes in non-NewtonMaximize related code, so there must be a performance bug somewhere to get a regression that large. I'll do some benchmarking/regression hunting tomorrow to figure this out.