@jeff-regier 2017-02-22
I ran `benchmark_seven_light_source.jl --profile` on master and this branch. The lines below show up in the profiler output only on this branch. They seem to account most of the difference in runtime.

```
    79 .../derivatives/elementwise.jl   197 (::ReverseDiff.##207#208{ForwardDiff.Dual{... 
    80 ./broadcast.jl                   117 _broadcast!(::ReverseDiff.##207#208{Forwar... 
    80 ./broadcast.jl                   172 broadcast!                                    
    80 ./broadcast.jl                   123 macro expansion                               
    80 ./broadcast.jl                   129 macro expansion                               
    80 ./simdloop.jl                     73 macro expansion                               
    80 .../derivatives/elementwise.jl   259 special_forward_exec!(::ReverseDiff.Specia... 
```