@jrevels 2017-02-22
Do you have non-truncated output (maybe post it in a gist)? This PR doesn't employ ReverseDiff in any new way at all vs. master - all the AD that is added here is ForwardDiff - so ReverseDiff shouldn't be involved in the regressions we're seeing.