@jrevels 2017-02-22
@jeff-regier I've fixed the performance regression I was seeing locally, let me know if that fixes it on your end. It was from an unnecessary change I made that stack-allocated the `ids` indices instead of heap-allocating them (i.e. nothing to do with AD or the substantive part of this PR). I'm still not certain why that caused a regression, but reverting that change fixed it.