@jeff-regier 2017-02-23
It's close now, about 5% difference. Looks like it could be due to GC time. Strange, since this branch has both fewer allocations and less memory allocated in total.

These runtimes are on a dedicated Cori Phase 1 node, with the version of Julia 0.6 that Andreas built, stored in his scratch directory. I'm running `benchmark_seven_light_sources.jl`.

master:
```
  5.207721 seconds (4.71 M allocations: 312.227 MiB, 1.20% gc time)
  5.175237 seconds (4.71 M allocations: 312.227 MiB, 1.19% gc time)
  5.203087 seconds (4.71 M allocations: 312.227 MiB, 1.21% gc time)
```

This branch:
```
  5.499711 seconds (4.26 M allocations: 183.664 MiB, 4.01% gc time)
  5.489853 seconds (4.26 M allocations: 183.664 MiB, 4.07% gc time)
  5.540679 seconds (4.26 M allocations: 183.664 MiB, 4.02% gc time)
```